refacted layers to nn methods
