+++ examples/mnist_tf/nn.py
@@ -1,5 +1,6 @@
 import numpy as np
 import tensorflow as tf
+from tensorflow.contrib.framework import add_arg_scope
 from tensorflow.contrib.framework.python.ops import add_arg_scope, arg_scope
 import math
 
@@ -9,3 +10,12 @@ def make_model(name, template, **kwargs):
     run = lambda *args, **kw: template(*args, **dict((k, v) for kws in (kw, kwargs) for k, v in kws.items()))
     return tf.make_template(name, run, unique_name_ = name)
 
+
+@add_arg_scope
+def conv2D(x, f, k, s=1, padding="same", activation=None):
+    return tf.layers.Conv2D(f, k, strides=s, padding=padding, activation=activation)(x)
+
+
+@add_arg_scope
+def dense(x, f, activation=None):
+    return tf.layers.dense(x, f, activation=activation)
